one listener for all sessions
