+++ completion-plugin/src/main/kotlin/org/jb/cce/highlighter/HighlightRange.kt
@@ -1,0 +1,3 @@
+package org.jb.cce.highlighter
+
+class HighlightRange(val begin: Int, val end: Int)
